check for h_addr_list added
Rev: src/modules/files/configure.in:1.5
Rev: src/modules/files/file.c:1.4
Rev: src/modules/files/file_machine.h.in:1.2
Rev: src/modules/spider/configure.in:1.3
Rev: src/modules/spider/sharedmem.c:1.2
Rev: src/modules/spider/spider.c:1.3
Rev: src/modules/spider/testsuite.in:1.2

+++ src/modules/spider/sharedmem.c
@@ -166,7 +166,7 @@ void shared_free(void *pointer)
 
 #else
 /* Should look for shmop as well, probably */
-void *shared_malloc(int size)
+void *shared_malloc(size_t size)
 {
 #ifdef SHARED_MALLOC_DEBUG
   fprintf(stderr, "fake_shared_malloc(%d)\n", size);
